+++ src/db_migrate.cpp
@@ -13,7 +13,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.1.58"};
+constexpr auto version = std::string_view{"0.1.59"};
 
 auto print_help() -> void
 {